requirements.txt
requests>=2.31.0
ollama>=0.1.0
openai-whisper>=20240101
faster-whisper>=1.0.3
websockets>=12.0
sounddevice>=0.4.6
numpy>=1.24.3
scipy>=1.11.4
soundfile>=0.12.1
mutagen>=1.46.0
musicbrainzngs>=0.7.1
python-dotenv>=1.0.0
pygame>=2.5.0
pydub>=0.25.1
simpleaudio>=1.0.4
edge-tts>=6.1.12
psutil>=5.9.8
wmi>=1.5.1
pycaw>=20240210
comtypes>=1.2.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
